% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMAQ_1km_figure.R
\name{CMAQ_1km_figure}
\alias{CMAQ_1km_figure}
\title{CMAQ_1km_figure}
\usage{
CMAQ_1km_figure(
  data = NULL,
  mapping = aes(),
  ...,
  drop_units = TRUE,
  gridlines = TRUE
)
}
\arguments{
\item{...}{passed to [ggplot2::ggplot()]}

\item{gridlines}{}
}
\value{
ggplot2 object
}
\description{
Create base for ggplot2 figure, using:
}
\details{
- [scale_CMAQ_1km()]
- [theme_CMAQ_1km()]
}
