% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMAQ_1km_geodata.R
\name{CMAQ_1km_geodata}
\alias{CMAQ_1km_geodata}
\title{CMAQ_1km_geodata}
\usage{
CMAQ_1km_geodata(...)
}
\arguments{
\item{...}{passed to [CMAQ_1km_raster()]}
}
\value{
sf object
}
\description{
Create a `sf` object with default 1km CRS and dimensions
}
