% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMAQ_1km_raster.R
\name{CMAQ_1km_raster}
\alias{CMAQ_1km_raster}
\title{CMAQ_1km_raster}
\usage{
CMAQ_1km_raster(dim = CMAQ_1km_dim(), crs = CMAQ_1km_crs())
}
\arguments{
\item{dim}{two-element numeric vector, as from [CMAQ_1km_dim()]}

\item{crs}{CRS object, as from [CMAQ_1km_crs()]}
}
\value{
[terra::rast] object
}
\description{
CMAQ_1km_raster
}
