% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMAQ_1km_row_from.R
\name{CMAQ_1km_row_from}
\alias{CMAQ_1km_row_from}
\title{CMAQ_1km_row_from}
\usage{
CMAQ_1km_row_from(cell, warn = TRUE)
}
\arguments{
\item{cell}{integer}

\item{warn}{logical}
}
\value{
integer
}
\description{
CMAQ_1km_row_from
}
