% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMAQ_1km_stars.R
\name{CMAQ_1km_stars}
\alias{CMAQ_1km_stars}
\alias{CMAQ_1km_stobj}
\title{CMAQ_1km_stars}
\usage{
CMAQ_1km_stars(...)
}
\arguments{
\item{...}{passed to [CMAQ_1km_raster()]}
}
\value{
stars object
}
\description{
Create a `stars` object with default 1km CRS and dimensions
}
