% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_CMAQ_1km.R
\name{coord_CMAQ_1km}
\alias{coord_CMAQ_1km}
\title{coord_CMAQ_1km}
\usage{
coord_CMAQ_1km(
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  crs = CMAQ_1km_crs(),
  datum = CMAQ_1km_crs(),
  ...
)
}
\arguments{
\item{xlim}{two-valued numeric vector}

\item{ylim}{two-valued numeric vector}

\item{crs}{CRS object, as from `CMAQ_1km_crs()`}

\item{datum}{CRS object, as from `CMAQ_1km_crs()`}

\item{...}{passed to [ggplot2::coord_sf()]}
}
\description{
For use with `ggplot()`, a projected `ccord_sf`
}
