% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_CMAQ_1km.R
\name{scale_CMAQ_1km}
\alias{scale_CMAQ_1km}
\title{scale_CMAQ_1km}
\usage{
scale_CMAQ_1km(
  names = c("CMAQ_col", "CMAQ_row"),
  ...,
  breaks = seq(0, 300, by = 20),
  expand = expansion(add = c(3, 3), mult = c(0, 0))
)
}
\arguments{
\item{...}{passed to [ggplot2::scale_x_continuous()] and [ggplot2::scale_x_continuous()]}

\item{breaks}{see [ggplot2::scale_continuous]}

\item{expand}{}
}
\value{
list of two scale objects, one for x and one for y
}
\description{
scale_CMAQ_1km
}
