% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_CMAQ_1km.R
\name{theme_CMAQ_1km}
\alias{theme_CMAQ_1km}
\title{theme_CMAQ_1km}
\usage{
theme_CMAQ_1km(..., gridlines = TRUE)
}
\arguments{
\item{...}{passed to [ggtools::theme_simple()]}

\item{gridlines}{logical}
}
\value{
theme object, as from [ggplot2::theme()]
}
\description{
theme_CMAQ_1km
}
