% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_CMAQ_1km_geometry.R
\name{with_CMAQ_1km_geometry}
\alias{with_CMAQ_1km_geometry}
\title{with_CMAQ_1km_geometry}
\usage{
with_CMAQ_1km_geometry(.data, by = CMAQ_id_vars(), relationship = "one-to-one")
}
\arguments{
\item{.data}{must contain columns given by [CMAQ_id_vars()]}

\item{relationship}{see [dplyr::left_join()]; default is "one-to-one"}
}
\value{
sf object with cell geometries
}
\description{
with_CMAQ_1km_geometry
}
